+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- [Library] Add color semantic tolens `colorBackgroundStatusNeutral`, `colorBackgroundStatusNeutralOnBackgroundEmphasized`
+- [Library] Add color semantic tokens `colorBackgroundStatusNeutral` and `OnBackgroundEmphasized` variants
 - [Library] Add typography semantic tokens for font letter spacing
 - [Library] Unit tests for multiple tokens
 - [Library] Add color semantic composite tokens embeding light and dark modes values